started a gui with tkinter
